+++ lib/mongoid/history.rb
@@ -19,12 +19,24 @@ module Mongoid
       attr_accessor :modifier_class_name
       attr_accessor :current_user_method
 
-      def disable(&_block)
+      def disable
+        original_flag = store[GLOBAL_TRACK_HISTORY_FLAG]
         store[GLOBAL_TRACK_HISTORY_FLAG] = false
-        yield
+        yield if block_given?
       ensure
-        store[GLOBAL_TRACK_HISTORY_FLAG] = true
+        store[GLOBAL_TRACK_HISTORY_FLAG] = original_flag if block_given?
       end
+
+      def enable
+        original_flag = store[GLOBAL_TRACK_HISTORY_FLAG]
+        store[GLOBAL_TRACK_HISTORY_FLAG] = true
+        yield if block_given?
+      ensure
+        store[GLOBAL_TRACK_HISTORY_FLAG] = original_flag if block_given?
+      end
+
+      alias disable! disable
+      alias enable! enable
 
       def enabled?
         store[GLOBAL_TRACK_HISTORY_FLAG] != false